Move plugin api types to iota-types
