+++ client/src/error.rs
@@ -26,7 +26,7 @@ pub enum Error {
     /// Block dtos error
     #[error("{0}")]
     #[serde(serialize_with = "display_string")]
-    ApiTypes(#[from] iota_types::api::error::Error),
+    ApiTypes(#[from] iota_types::api::core::error::Error),
     /// Blake2b256 Error
     #[error("{0}")]
     Blake2b256(&'static str),
@@ -247,18 +247,10 @@ pub enum Error {
     #[error("can't burn and transition an output at the same time, chain ID: {0}")]
     BurnAndTransition(ChainId),
 
-    //////////////////////////////////////////////////////////////////////
-    // Participation
-    //////////////////////////////////////////////////////////////////////
-    /// Invalid participations error
-    #[cfg(feature = "participation")]
-    #[error("invalid participations")]
-    InvalidParticipations,
-    /// IO error
-    #[cfg(feature = "participation")]
-    #[error("`{0}`")]
+    /// Participation error
+    #[error("{0}")]
     #[serde(serialize_with = "display_string")]
-    Io(#[from] std::io::Error),
+    Participation(#[from] iota_types::api::plugins::participation::Error),
 
     //////////////////////////////////////////////////////////////////////
     // Ledger Nano
