+++ client/src/node_manager/mod.rs
@@ -15,7 +15,7 @@ use std::{
     time::Duration,
 };
 
-use iota_types::api::response::InfoResponse;
+use iota_types::api::core::response::InfoResponse;
 use serde_json::Value;
 
 use self::{http_client::HttpClient, node::Node};
