+++ types/Cargo.toml
@@ -22,6 +22,7 @@ bech32 = { version = "0.9.1", default-features = false }
 bitflags = { version = "1.3.2", default-features = false }
 bytemuck = { version = "1.13.0", default-features = false }
 derive_more = { version = "0.99.17", default-features = false, features = [ "from", "as_ref", "deref", "deref_mut" ] }
+getset = { version = "0.1.2", default-features = false, optional = true }
 hashbrown = { version = "0.13.2", default-features = false, features = [ "ahash", "inline-more", "serde" ] }
 hex = { version = "0.4.3", default-features = false, features = [ "alloc" ] }
 iota-crypto = { version = "0.15.3", default-features = false, features = [ "ed25519", "blake2b", "ternary_encoding" ] }
@@ -34,6 +35,7 @@ rand = { version = "0.8.5", default-features = false, features = [ "min_const_ge
 serde = { version = "1.0.152", default-features = false, features = [ "derive" ], optional = true }
 serde-big-array = { version = "0.4.1", default-features = false, optional = true }
 serde_json = { version = "1.0.91", default-features = false, features = [ "alloc" ], optional = true }
+serde_repr = { version = "0.1.10", default-features = false, optional = true }
 thiserror = { version = "1.0.38", default-features = false }
 
 [dev-dependencies]
@@ -42,7 +44,7 @@ num_cpus = { version = "1.15.0", default-features = false }
 [features]
 default = [ "api", "block", "std" ]
 
-api = [ "block", "dto" ]
+api = [ "block", "dto", "getset", "serde_repr" ]
 block = [  ]
 dto = [ "serde", "dep:serde_json" ]
 rand = [ "dep:rand", "std" ]
