+++ types/src/api/core/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2020-2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Core API-related types like responses and DTOs.
+
+pub mod dto;
+pub mod error;
+pub mod response;
