+++ types/src/api/plugins/participation/mod.rs
@@ -1,0 +1,20 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Node participation API.
+//! https://github.com/iota-community/treasury/blob/main/specifications/hornet-participation-plugin.md
+
+pub mod responses;
+pub mod types;
+
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    /// Invalid participations error
+    #[error("invalid participations")]
+    InvalidParticipations,
+    /// IO error
+    #[error("`{0}`")]
+    Io(#[from] std::io::Error),
+}
