+++ types/tests/mod.rs
@@ -2,4 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod address;
+mod api;
 mod input;
